+++ src/pages/pupils/beta/programmes/[programmeSlug]/units/[unitSlug]/lessons.tsx
@@ -1,11 +1,10 @@
 import { GetStaticProps, GetStaticPropsResult } from "next";
-import Link from "next/link";
 
 import curriculumApi2023 from "@/node-lib/curriculum-api-2023";
 import { LessonListingBrowseData } from "@/node-lib/curriculum-api-2023/queries/pupilLessonListing/pupilLessonListing.schema";
 import getPageProps from "@/node-lib/getPageProps";
-import { resolveOakHref } from "@/common-lib/urls";
 import { getStaticPaths as getStaticPathsTemplate } from "@/pages-helpers/get-static-paths";
+import { PupilViewsLessonListing } from "@/components/PupilViews/PupilLessonListing/PupilLessonListing.view";
 
 type PupilLessonListingURLParams = {
   programmeSlug: string;
@@ -16,12 +15,10 @@ export type LessonListingPageProps = {
   curriculumData: LessonListingBrowseData;
 };
 
-export const getStaticPaths =
-  getStaticPathsTemplate<PupilLessonListingURLParams>;
-
 const PupilLessonListingPage = ({ curriculumData }: LessonListingPageProps) => {
   const unitData = curriculumData[0]?.unitData;
   const programmeFields = curriculumData[0]?.programmeFields;
+  const programmeSlug = curriculumData[0]?.programmeSlug;
 
   const orderedCurriculumData = curriculumData.sort((a, b) => {
     const aLessonOrder = a.supplementaryData?.orderInUnit;
@@ -29,33 +26,25 @@ const PupilLessonListingPage = ({ curriculumData }: LessonListingPageProps) => {
     return aLessonOrder - bLessonOrder;
   });
 
+  if (
+    unitData === undefined ||
+    programmeFields === undefined ||
+    programmeSlug === undefined
+  ) {
+    throw new Error("unitData or programmeFields is undefined");
+  }
   return (
-    <div>
-      <h1>{unitData?.title}</h1>
-      <h2>{programmeFields?.subject}</h2>
-      <h3>{programmeFields?.yearDescription}</h3>
-      <ol>
-        {orderedCurriculumData.map((lesson) => {
-          const lessonData = lesson.lessonData;
-          return (
-            <li key={lesson.lessonSlug}>
-              <Link
-                href={resolveOakHref({
-                  page: "pupil-lesson",
-                  lessonSlug: lesson.lessonSlug,
-                  programmeSlug: lesson.programmeSlug,
-                  unitSlug: lesson.unitSlug,
-                })}
-              >
-                {lessonData?.title}
-              </Link>
-            </li>
-          );
-        })}
-      </ol>
-    </div>
+    <PupilViewsLessonListing
+      unitData={unitData}
+      programmeFields={programmeFields}
+      orderedCurriculumData={orderedCurriculumData}
+      programmeSlug={programmeSlug}
+    />
   );
 };
+
+export const getStaticPaths =
+  getStaticPathsTemplate<PupilLessonListingURLParams>;
 
 export const getStaticProps: GetStaticProps<
   LessonListingPageProps,